requirements.txt
aiohttp>=3.11.16
azure-identity>=1.21.0
azure-storage-queue>=12.12.0
pytest>=8.3.5